Support more licenses in COPYING/LICENSE
Instead of trying to put a bunch of licenses in one file with jinja2 if
statements, keep them in separate files, and copy the right one to the
specified location in the post generation hook.

Also re-organize post_gen_project.py into functions and use the standard
main() calling convention.

Change-Id: I28ad84b7985dc5390058ada48809b739bfe0f266