Adjust the version of devDependencies set to 2025-10-08 standards
Switching to PHP 8.1+ means a number of changes to the phpunit.xml.dist
file.

Taking the opportunity to also move canonical phpunit testing from
tests/ to tests/phpunit/, so that libraries can support multiple testing
tools in one tree (as many do), and to be more canonical with other
libraries.

Bug: T395743
Change-Id: I84afb64bb1eac4f88e1b4b359cb093e00dd077d2